pyproject.toml
[tool.poetry]
name = "bookwarms-backend"
version = "0.1.0"
description = ""
authors = ["KangseokOh"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.4"
djangorestframework = "^3.15.1"
mysql-connector-python = "^8.3.0"
django-cors-headers = "^4.3.1"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
